refactor(docs): extract playground-state types leaf (#462)
Move diagnostic/snapshot/candidate/state types and PLAYGROUND_MAX_UNDO_SNAPSHOTS
into playground-state-types.ts so draft-validate (and other pure modules) can
import types without a cycle against the state machine.

playground-state re-exports the leaf for stable public import paths.

Closes #462.
